requirements.txt
numpy
opencv-python
scikit-learn
matplotlib
kagglehub
tensorflow
pyyaml